WAITP-1219 Update file arrangement
