+++ packages/lesmis/src/components/Visuals/ListVisual/ColorCircle.js
@@ -6,7 +6,7 @@ import React from 'react';
 import styled from 'styled-components';
 import { lighten } from '@material-ui/core/styles';
 import PropTypes from 'prop-types';
-import { Tooltip } from '../Tooltip';
+import { Tooltip } from '@tupaia/ui-components';
 
 const CircleCell = styled.div`
   display: flex;
