+++ packages/lesmis/src/components/Visuals/ListVisual/index.js
@@ -1,0 +1,1 @@
+export { ListVisual } from './ListVisual';
