+++ packages/ui-chart-components/src/utils/utils.js
@@ -6,6 +6,7 @@
 import moment from 'moment';
 import { useTheme } from '@material-ui/core/styles';
 import { GRANULARITY_CONFIG } from '@tupaia/utils';
+import { CHART_TYPES } from '../constants';
 
 export const isMobile = () => process.env.REACT_APP_APP_TYPE === 'mobile';
 
@@ -21,4 +22,29 @@ export const isDataKey = key =>
 export const getContrastTextColor = () => {
   const theme = useTheme();
   return theme.palette.type === 'light' ? theme.palette.text.secondary : 'white';
+};
+
+export const getIsChartData = ({ chartType, data }) => {
+  // If all segments of a pie chart are "0", display the no data message
+  if (chartType === CHART_TYPES.PIE && data && data.every(segment => segment.value === 0)) {
+    return false;
+  }
+
+  return data && data.length > 0;
+};
+
+export const getNoDataString = ({ noDataMessage, source, startDate, endDate }) => {
+  if (noDataMessage) {
+    return noDataMessage;
+  }
+
+  if (source === 'mSupply') {
+    return 'Requires mSupply';
+  }
+
+  if (startDate && endDate) {
+    return `No data for ${startDate} to ${endDate}`;
+  }
+
+  return 'No data for selected dates';
 };
