+++ packages/utils/src/__tests__/period/periodGranularities.test.js
@@ -5,7 +5,7 @@
 
 import moment from 'moment';
 import sinon from 'sinon';
-import { roundStartEndDates, getDefaultDates, getLimits } from '../../period/chartGranularities';
+import { roundStartEndDates, getDefaultDates, getLimits } from '../../period/periodGranularities';
 
 const DEFAULT_NOW_TIMESTAMP = 1549360800000; // 2019-02-05T10:00:00.000Z
