+++ packages/web-frontend/src/components/View/View.js
@@ -6,7 +6,8 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import CircularProgress from 'material-ui/CircularProgress';
-import { periodToMoment, CHART_TYPES } from '@tupaia/utils';
+import { periodToMoment } from '@tupaia/utils';
+import { CHART_TYPES } from '@tupaia/ui-chart-components';
 import { VIEW_STYLES } from '../../styles';
 import { NoDataMessage } from './NoDataMessage';
 import { VIEW_CONTENT_SHAPE } from './propTypes';
